test(sidecar): pin Do decode and inventory-failure behavior in its own package
